sleep_mode: Allow setting a 0 sleep time
Allow setting a 0 minute sleep time, so that the device goes to sleep
instantly. This is to address the alarm on lock issue [1] where the car
will activate an alarm if OBD2 packets are sent while the car is locked.

If we go to sleep as soon as the battery stops charging we then ensure
that the traffic will stop by the time the driver locks the car.

This might still have issues charging a locked EV, where the car will
charge the 12V battery while charging the car, which will wakeup the
WiCAN causing the alarm to activate. Limited testing of the ID. Buzz
doesn't do this (at least not right away). Combining with [2] might help
address this in the long run.

Obviously this also requires no periodic wakeup, which means that no
information will be avaliable from the ECU when the car isn't running.
Which is less then ideal, but at least the alarm doesn't go off.

[2] would be a great improvement as well, but has the risk that an
accidental lock of the car will set off the alarm, which is very
annoying.

1: https://github.com/meatpiHQ/wican-fw/issues/506
2: https://github.com/meatpiHQ/wican-fw/issues/507

Signed-off-by: Alistair Francis <alistair@alistair23.me>