components/autopid: Only probe ECU if we are connected to something
If we aren't connected to WiFi, BLE or MQTT then there is no point in
polling the car for information, as we have nothing to do with the info.
This way if nothing is connected we don't poll the car, this hopefully
avoids setting off the car alarm as long as the user disconnects from
all networks before locking the car.

It won't fix all alarm scenarios, but it helps mitigate a few of them.
Espicially considering most people leave there car unlocked at home
(WiFi) and take there phone with them away from the car (BLE).
Hopefully this means the device won't poll and set off the alarm when
locked when out, but will still work fine at home.

Note that in the future this should be updated to check if persistient
logging is enabled as well.

Resolves: https://github.com/meatpiHQ/wican-fw/issues/507
Signed-off-by: Alistair Francis <alistair@alistair23.me>